Use new instantiators in ExtensionManager
